subi al repositorio

+++ templates_c/4098d78bc5fd5ca5837627d485104016ae9aa237_0.file.productsDetail.tpl.php
@@ -1,18 +1,18 @@
 <?php
-/* Smarty version 4.2.1, created on 2022-10-14 01:04:14
+/* Smarty version 4.2.1, created on 2022-10-15 22:15:54
   from 'C:\xampp\htdocs\web2\tpeweb\templates\productsDetail.tpl' */
 
 /* @var Smarty_Internal_Template $_smarty_tpl */
 if ($_smarty_tpl->_decodeProperties($_smarty_tpl, array (
   'version' => '4.2.1',
-  'unifunc' => 'content_6348996ecdc489_57539693',
+  'unifunc' => 'content_634b14fa9636a9_85057704',
   'has_nocache_code' => false,
   'file_dependency' => 
   array (
     '4098d78bc5fd5ca5837627d485104016ae9aa237' => 
     array (
       0 => 'C:\\xampp\\htdocs\\web2\\tpeweb\\templates\\productsDetail.tpl',
-      1 => 1665686759,
+      1 => 1665864953,
       2 => 'file',
     ),
   ),
@@ -22,7 +22,7 @@ if ($_smarty_tpl->_decodeProperties($_smarty_tpl, array (
     'file:footer.tpl' => 1,
   ),
 ),false)) {
-function content_6348996ecdc489_57539693 (Smarty_Internal_Template $_smarty_tpl) {
+function content_634b14fa9636a9_85057704 (Smarty_Internal_Template $_smarty_tpl) {
 $_smarty_tpl->_subTemplateRender("file:header.tpl", $_smarty_tpl->cache_id, $_smarty_tpl->compile_id, 0, $_smarty_tpl->cache_lifetime, array(), 0, false);
 ?>
 
@@ -45,10 +45,12 @@ $_smarty_tpl->tpl_vars['product']->do_else = false;
                 <p class="card-text">$<?php echo $_smarty_tpl->tpl_vars['product']->value->precio;?>
  - <?php echo $_smarty_tpl->tpl_vars['product']->value->talle;?>
 </p>
-                <a href='delete/<?php echo $_smarty_tpl->tpl_vars['product']->value->id;?>
+                <?php if ((isset($_SESSION['IS_LOGGED']))) {?>
+                    <a href='delete/<?php echo $_smarty_tpl->tpl_vars['product']->value->id;?>
 ' type='button' class='btn btn-danger'>Borrar</a>
-                <a href='edit/<?php echo $_smarty_tpl->tpl_vars['product']->value->id;?>
+                    <a href='edit/<?php echo $_smarty_tpl->tpl_vars['product']->value->id;?>
 '  type='button' class='btn btn-warning'>Editar</a>     
+                <?php }?> 
             </div>
         </div>
     <?php
